added date class and refactored affected methods and tests
