+++ spec/spec_helper.rb
@@ -10,3 +10,5 @@ Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
 
 require_relative '../lib/reservation'
 require_relative '../lib/reservation_hub'
+require_relative '../lib/room_block'
+require_relative '../lib/date_range'
